Extract pagination resolvers to util and share variable-reference schema
- Move resolveLimitInput / resolveOffsetInput into a record-crud util
- Extract workflowVariableReferenceSchema to a shared schema and reuse it in
  find-records and send-email settings
