+++ packages/twenty-shared/src/workflow/schemas/workflow-variable-reference-schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const workflowVariableReferenceSchema = z
+  .string()
+  .regex(
+    /^{{[^{}]+}}$/,
+    'Expected a workflow variable reference like {{stepId.path}}',
+  );
